fix: address PRRT_kwDOSJAM6s6HsXLk — split Docker pull detection out of pr_monitor
pr_monitor.py had grown to 1,540 lines, breaching the 1,500-line
first-party guardrail enforced by
test_core_decomposition_maintainability.py::test_first_party_code_files_stay_under_line_limit.

Extract the six Docker pull / OCI registry timeout detection helpers
(_is_docker_pull_failure_line, _image_pull_failure_is_corroborated,
_docker_pull_command_succeeded, _forward_detail_ref_matches_pull,
_evidence_line_is_permanent_pull_failure, _log_shows_docker_registry_timeout)
and their ~165-line constant block into a new, standalone module
_docker_pull_detection.py.  These functions operate only on primitive
types (str, list[str], int) and carry no domain-type dependencies, so
they form a clean, independently-importable unit.

pr_monitor.py imports _log_shows_docker_registry_timeout from the new
module, preserving backward-compatible access for tests that import it
from awf.runtime.pr_monitor.  pr_monitor.py drops from 1,540 to 1,028
lines; _docker_pull_detection.py is 527 lines.
